Snek in C/SDL2
This is not the best possible implementation and some code improvements
could be done. Linked list implementation forbids placing the same on
more than one list. Audio playback opens a global resource in a way
that allows for that resource to be opened more than once. The audio
system is also tied into the game logic even more so than it is in
the C++ version - there's a direct audio module function call with
dereferencing a global SnekcAudio pointer.

That said I'm more than surprised how clean the C version turned out to
be. With the removal of C++ boiler-plate incantations, what remains is
almost pure "business logic".