Amiga: localized tile clears are mandatory to prevent flashing
The global screen clear every frame doesn't cut it with this
approach to Amiga with Intuition. I'm most certainly missing
something here. For now the screen flashing can be avoided
by only clearing a small portion of the screen that actually
needs to be cleared: the final part of the tail and the score
display. Fullscreen clear occurs only when the level is reset.

I might want to reapproach Amiga differently but for now
I lack the necessary knowledge to get this done right.